+++ client/bindings/nodejs/types/bridge/client.ts
@@ -4,8 +4,8 @@ import type { IGenerateAddressesOptions } from '../generateAddressesOptions';
 import type { IBuildBlockOptions } from '../buildBlockOptions';
 import type { BlockId } from '../blockId';
 import type {
-    IInputSigningData,
     IPreparedTransactionData,
+    IBip32Chain,
 } from '../preparedTransactionData';
 import type {
     AliasQueryParameter,
@@ -166,10 +166,8 @@ export interface __SignatureUnlockMessage__ {
     name: 'signatureUnlock';
     data: {
         secretManager: SecretManager;
-        inputSigningData: IInputSigningData;
         transactionEssenceHash: Array<number>;
-        //TODO: Expose `RemainderData`.
-        remainderData: undefined;
+        chain: IBip32Chain;
     };
 }