+++ client/src/error.rs
@@ -53,6 +53,14 @@ pub enum Error {
     /// Invalid BIP32 chain data
     #[error("invalid BIP32 chain data")]
     InvalidBIP32ChainData,
+    /// Invalid bech32 HRP, should match the one from the used network
+    #[error("invalid bech32 hrp for the connected network: {provided}, expected: {expected}")]
+    InvalidBech32Hrp {
+        /// The bech32 human readable part from the provided address.
+        provided: String,
+        /// The expected bech32 human readable part.
+        expected: String,
+    },
     /// Invalid mnemonic error
     #[error("invalid mnemonic {0}")]
     InvalidMnemonic(String),
@@ -181,6 +189,9 @@ pub enum Error {
     /// Input selection error.
     #[error("{0}")]
     InputSelection(#[from] InputSelectionError),
+    /// Missing BIP32 chain to sign with.
+    #[error("missing BIP32 chain to sign with")]
+    MissingBip32Chain,
 
     /// Participation error
     #[cfg(feature = "participation")]